Run tests using real config
There are lots of situations where we're going to need to depend on the
existence of real third party plugins during testing. At some point I'll
add support for a bundled testing config, but this is a hack that works
for now, since I really just want to keep moving forwards.